requirements.txt
torch>=2.0.0
tqdm>=4.65.0
h5py>=3.9.0
numpy>=1.24.0
PyYAML>=6.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0
jiwer
